win7 v0.9. rfrb preAlpha, GNULicense added

+++ win7/include/thumbWriter.h
@@ -18,7 +18,7 @@ class ThumbWriter
 		bool writeXml;
 		bool edbOnly;
 		bool isExtracted;
-		size_t offset;
+		int offset;
 
 	public:
 		ThumbWriter(std::string in, std::string out);
@@ -27,7 +27,7 @@ class ThumbWriter
 		void setXml(bool xml);
 		void setEdb(bool edb);
 		void setExtracted(bool ex);
-		void setOffset(size_t off);
+		void setOffset(int off);
 
 		void writeThumbs();
